Biphasic got three controls (including polarity)
TIM2 is coupled to TIM1 now. TIM2 outputs a logic signal for switching
polarity of the electrode wrt case based on TIM1 timings. TIM2 DMA
requests update TIM2's OC1 register; when a switch is desired, set OC1
to be low (~10-20 to carry over the TIM1 pulse and allow the physical
pulse to die before switching); when no switch is desired, set OC1 to
0xFFFF. In current configuration, polarity switches 50us after the each
pulse (positive or negative) dies.